+++ tests/conftest.py
@@ -7,3 +7,4 @@ import pytest  # noqa  # pylint: disable=unused-import
 from tests.fixtures.mock_setup import *  # noqa: F401
 from tests.fixtures.qcodes_utilities import *  # noqa: F401
 from tests.fixtures.schedule import *  # noqa: F401
+from tests.scheduler.backends.qblox.fixtures.hardware_config import *  # noqa: F401