+++ tests/fixtures/mock_setup.py
@@ -10,6 +10,7 @@ from typing import Any, Dict, List, Union
 import pytest
 from qcodes import Instrument
 from quantify_core.data.handling import get_datadir, set_datadir
+
 from quantify_scheduler.device_under_test.mock_setup import (
     set_standard_params_basic_nv,
     set_standard_params_transmon,
@@ -194,6 +195,24 @@ def compile_config_basic_transmon_qblox_hardware(
     mock_setup = mock_setup_basic_transmon_with_standard_params
     mock_setup["quantum_device"].hardware_config(QBLOX_HARDWARE_MAPPING)
     mock_setup["quantum_device"].hardware_options(QBLOX_HARDWARE_OPTIONS)
+
+    yield mock_setup["quantum_device"].generate_compilation_config()
+
+
+@pytest.fixture(scope="function", autouse=False)
+def compile_config_basic_transmon_qblox_hardware_pulsar(
+    mock_setup_basic_transmon_with_standard_params,
+    hardware_cfg_pulsar,
+):
+    """
+    A config for a quantum device with 5 transmon qubits connected in a star
+    configuration controlled using Qblox Pulsars.
+    """
+    # N.B. how this fixture produces the hardware config will change in the future
+    # as we separate the config up into a more fine grained config. For now it uses
+    # the old JSON files to load settings from.
+    mock_setup = mock_setup_basic_transmon_with_standard_params
+    mock_setup["quantum_device"].hardware_config(hardware_cfg_pulsar)
 
     yield mock_setup["quantum_device"].generate_compilation_config()