+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Gettables - Clean up code syntax (!638)
 - Schedulables - Make name uniqueness check more efficient and readable. (!631)
 - Compilation - `CompilationNode`s take the full `CompilationConfig` as input (!615, #405)
+- Tests - Move Qblox Pulsar entries from `qblox_test_mapping.json` to pytest fixtures (!632)
 - Visualization - Introduce the `x_range` keyword for the matplotlib backend in `Schedule.plot_pulse_diagram`. This will cut off any points outside the given range when creating the plot. This can be used to reduce memory usage when plotting a small section of a long pulse sequence (!629).
 - Schedules - Add `nv_dark_esr_sched_nco` spectroscopy schedule using SetClockFrequency Operation to sweep the NCO frequency (!639)
 - ZI LabOne backend - Return datasets from UHFQA instrument coordinator component (which fixes the broken backend) (#410, !623).